requirements.txt
torch==1.8.1
torchvision==0.9.1
matplotlib==3.2.2
numpy==1.21.6
scipy==1.7.3
scikit-learn==1.0.2
seaborn==0.12.2
tqdm==4.65.0
requests==2.31.0
pandas==0.25
pillow==4.1.1
six==1.5
absl-py